+++ packages/runtime/src/types.ts
@@ -393,3 +393,8 @@ export interface SCParams {
 export interface ReplaceParams {
 	[key: string]: string;
 }
+
+export enum Base64Encoding {
+	URL = 0,
+	STD = 1,
+}