+++ packages/runtime/test/src/parser/parser.ts
@@ -22,6 +22,7 @@ import {
 	Arg,
 	Assert,
 	Balance,
+	Base64Decode,
 	BitLen,
 	BitwiseAnd,
 	BitwiseNot,
@@ -2266,7 +2267,14 @@ describe("Parser", function () {
 			);
 		});
 
-		it("Should failed if declare pragma greater than 6", () => {
+		it("Supported pragma version 7", () => {
+			const fileWithPragmav7 = "test-pragma-v7.teal";
+			assert.doesNotThrow(() =>
+				parser(getProgram(fileWithPragmav7), ExecutionMode.SIGNATURE, interpreter)
+			);
+		});
+
+		it("Should fail if declare pragma greater than 7", () => {
 			const fileWithPragmaInvalid = "test-pragma-invalid.teal";
 			expectRuntimeError(
 				() => parser(getProgram(fileWithPragmaInvalid), ExecutionMode.SIGNATURE, interpreter),
@@ -2582,6 +2590,17 @@ describe("Parser", function () {
 			];
 
 			assert.deepEqual(res, expected);
+		});
+
+		it("should return correct opcode list for `teal v7`", async () => {
+			const file = "teal-v7.teal";
+			const res = parser(getProgram(file), ExecutionMode.APPLICATION, interpreter);
+			const expected = [
+				new Pragma(["version", "7"], 1, interpreter),
+				new Base64Decode(["URLEncoding"], 2),
+			];
+
+			assert.deepEqual(expected, res);
 		});
 	});